pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-auth-service"
version = "0.1.0"
description = "Production-minded authentication service with FastAPI"
requires-python = ">=3.11,<3.13"

# Core runtime dependencies (constrained for stability)
dependencies = [
  "fastapi>=0.110,<0.112",
  "uvicorn>=0.29,<0.31",
  "sqlalchemy>=2.0.29,<2.1",
  "psycopg[binary]>=3.1.18,<3.2",
  "alembic>=1.13.1,<1.14",
  "pydantic>=2.6,<2.8",
  "pydantic-settings>=2.2,<2.4",
  "email-validator>=2.1,<3",
  "passlib[argon2]>=1.7.4,<1.8",
  "PyJWT[crypto]>=2.8,<2.10",
  "structlog>=24.1,<24.3",
  "python-dotenv>=1.0.1,<1.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1,<8.3",
  "pytest-asyncio>=0.23,<0.25",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.3,<0.5",
  "mypy>=1.9,<1.12",
  "types-requests>=2.31,<2.33",
  "httpx>=0.27,<0.28",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app", "app.*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
